# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Brain AI System API
  description: |
    A comprehensive API for the Brain AI system, providing access to character-level prediction,
    segment discovery, memory management, concept graphs, insight extraction, simulation capabilities,
    and advanced query/export functionality.
    
    ## Authentication
    
    The API supports two authentication methods:
    - **JWT Tokens**: Bearer tokens for session-based authentication
    - **API Keys**: Long-lived keys for programmatic access
    
    ## Rate Limiting
    
    Rate limits are enforced based on user roles:
    - **Admin**: 1000 requests/minute
    - **Developer**: 500 requests/minute  
    - **Analyst**: 200 requests/minute
    - **User**: 100 requests/minute
    - **Guest**: 20 requests/minute
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns detailed error information:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `429` - Too Many Requests (rate limit exceeded)
    - `500` - Internal Server Error
    
  version: 0.1.0
  contact:
    name: Brain AI Team
    email: admin@brain-ai.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.brain-ai.com/v1
    description: Production server
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /segment:
    post:
      tags:
        - Text Processing
      summary: Segment text into processable units
      description: Break down input text into segments using BPE and pattern discovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  example: "The quick brown fox jumps over the lazy dog"
                max_segments:
                  type: integer
                  minimum: 1
                  maximum: 1000
                  default: 100
                  description: Maximum number of segments to return
      responses:
        '200':
          description: Text segmented successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  segments:
                    type: array
                    items:
                      $ref: '#/components/schemas/TextSegment'
                  processing_time_ms:
                    type: integer
                  metadata:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /learn:
    post:
      tags:
        - Learning
      summary: Learn new information
      description: Add new information to the system's knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - information
              properties:
                information:
                  type: string
                  example: "User prefers coffee in the morning"
                priority:
                  type: string
                  enum: [low, medium, high, critical]
                  default: medium
                context:
                  type: object
                  description: Additional context metadata
      responses:
        '200':
          description: Information learned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  memory_id:
                    type: string
                  concepts_created:
                    type: integer
                  rules_generated:
                    type: integer
                  processing_time_ms:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /simulate:
    post:
      tags:
        - Simulation
      summary: Run predictive simulation
      description: Execute a simulation based on a scenario description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - scenario
              properties:
                scenario:
                  type: string
                  example: "What happens if the user wakes up early?"
                max_steps:
                  type: integer
                  minimum: 1
                  maximum: 100
                  default: 10
                confidence_threshold:
                  type: number
                  minimum: 0.0
                  maximum: 1.0
                  default: 0.5
      responses:
        '200':
          description: Simulation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /query/memory:
    post:
      tags:
        - Querying
      summary: Query memory system
      description: Search through the system's memory using natural language or structured queries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  example: "coffee preferences"
                limit:
                  type: integer
                  minimum: 1
                  maximum: 1000
                  default: 50
                memory_types:
                  type: array
                  items:
                    type: string
                    enum: [working, episodic, semantic]
                  example: ["episodic", "semantic"]
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemoryResult'
                  total_results:
                    type: integer
                  query_time_ms:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /export:
    post:
      tags:
        - Data Export
      summary: Export system data
      description: Export concepts, memories, or rules in various formats
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - format
              properties:
                format:
                  type: string
                  enum: [json, csv, graph]
                  example: "json"
                data_types:
                  type: array
                  items:
                    type: string
                    enum: [concepts, memories, rules]
                  example: ["concepts", "memories"]
                include_metadata:
                  type: boolean
                  default: true
                filters:
                  type: object
                  description: Optional filters to apply to exported data
      responses:
        '200':
          description: Data exported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResult'
            text/csv:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # System Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    TextSegment:
      type: object
      properties:
        text:
          type: string
          description: The segmented text content
        start:
          type: integer
          description: Start position in original text
        end:
          type: integer
          description: End position in original text
        confidence:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Confidence score for the segmentation
        segment_type:
          type: string
          description: Type of segment (character, word, phrase, etc.)
    SimulationResult:
      type: object
      properties:
        outcome:
          type: string
          description: Simulation outcome description
        confidence:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Confidence in the prediction
        steps:
          type: integer
          description: Number of simulation steps taken
        execution_time_ms:
          type: integer
          description: Execution time in milliseconds
        metadata:
          type: object
          description: Additional simulation metadata
    MemoryResult:
      type: object
      properties:
        content:
          type: string
          description: Memory content
        memory_type:
          type: string
          enum: [working, episodic, semantic]
        relevance:
          type: number
          minimum: 0.0
          maximum: 1.0
          description: Relevance score to the query
        timestamp:
          type: string
          format: date-time
        importance:
          type: string
          enum: [low, medium, high, critical]
...
    ExportResult:
      type: object
      properties:
        format:
          type: string
          enum: [json, csv, graph]
        data:
          type: string
          description: Exported data (base64 encoded for binary formats)
        metadata:
          type: object
          description: Export metadata
        stats:
          type: object
          description: Export statistics
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type or code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        request_id:
          type: string
          description: Unique request identifier for tracking
  responses:
    BadRequest:
      description: Bad request - validation error or malformed input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input parameters"
            details:
              field: "text"
              issue: "Text cannot be empty"
            request_id: "req_123456789"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Authentication required"
            request_id: "req_123456789"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "forbidden"
            message: "Insufficient permissions for this operation"
            request_id: "req_123456789"
...